requirements.txt
streamlit
cohere
PyPDF2
pdf2image
pytesseract
faiss-cpu
pillow
numpy